Revert "bugfix, set temp value to long"
This reverts commit e82208d2a932b2e862786794b4e321bfea39b607.

+++ polling-service/main.py
@@ -19,7 +19,7 @@ def createTempReading(temp):
             },
             "time": timestampToDatetime(time.time()),
             "fields" : {
-                "value": long(temp)
+                "value": temp
             }
         }
     ]
